feat: highlight secondary nav items on scroll
