+++ src/components/NavTree/index.ts
@@ -1,2 +1,3 @@
 export { NavTree } from "./NavTree";
-export type { NavItemConfig, NavTreeProps } from "./NavTree";
+export type { NavTreeProps } from "./NavTree";
+export type { GetIsActiveItemFunction, NavItemConfig } from "./types";
