+++ src/views/Package/PackageDocs.tsx
@@ -1,5 +1,5 @@
 import { Box, Flex, Grid } from "@chakra-ui/react";
-import { FunctionComponent, useEffect, useMemo } from "react";
+import { FunctionComponent, useEffect } from "react";
 import {
   Route,
   Switch,
@@ -7,20 +7,15 @@ import {
   useLocation,
   Redirect,
 } from "react-router-dom";
-import { NavTree } from "../../components/NavTree";
 import { ChooseSubmodule } from "./ChooseSubmodule";
-import {
-  API_URL_RESOURCE,
-  README_ITEM_ID,
-  PACKAGE_ANALYTICS,
-} from "./constants";
+import { API_URL_RESOURCE } from "./constants";
 import { NavDrawer } from "./NavDrawer";
 import { PackageReadme } from "./PackageReadme";
 import { usePackageState } from "./PackageState";
 import { PackageTypeDocs } from "./PackageTypeDocs";
+import { PrimaryDocNavigation } from "./PrimaryDocNavigation";
+import { SecondaryDocNavigation } from "./SecondaryDocNavigation";
 import { StickyNavContainer } from "./StickyNavContainer";
-import { useSectionItems } from "./useSectionItems";
-import { isApiPath, MenuItem } from "./util";
 
 // We want the nav to be sticky, but it should account for the sticky heading as well, which is 72px
 const TOP_OFFSET = "4.5rem";
@@ -45,25 +40,9 @@ const SubmoduleSelector: FunctionComponent = () => {
 
 export const PackageDocs: FunctionComponent = () => {
   const { path } = useRouteMatch();
-  const { menuItems, markdownDocs } = usePackageState();
-  const sectionItems = useSectionItems();
+  const { markdownDocs } = usePackageState();
 
   const { hash, pathname, search } = useLocation();
-
-  const primaryNavItems = useMemo(() => {
-    return menuItems.reduce((items, item, i) => {
-      // Omit README children which will be displayed on secondary nav
-      if (i === 0 && item?.id === README_ITEM_ID) {
-        const { children, ...readme } = item;
-
-        items.push({ ...readme, children: [] });
-      } else {
-        items.push(item);
-      }
-
-      return items;
-    }, [] as MenuItem[]);
-  }, [menuItems]);
 
   useEffect(() => {
     window.requestAnimationFrame(() => {
@@ -72,9 +51,6 @@ export const PackageDocs: FunctionComponent = () => {
           `[data-heading-id="${hash}"]`
         ) as HTMLElement;
 
-        target?.scrollIntoView(true);
-      } else if (isApiPath(pathname)) {
-        const target = document.getElementById(DOCS_ROOT_ID) as HTMLElement;
         target?.scrollIntoView(true);
       } else {
         window.scrollTo(0, 0);
@@ -104,10 +80,7 @@ export const PackageDocs: FunctionComponent = () => {
       >
         <SubmoduleSelector />
         <Box overflowY="auto" py={4}>
-          <NavTree
-            data-event={PACKAGE_ANALYTICS.SCOPE}
-            items={primaryNavItems}
-          />
+          <PrimaryDocNavigation />
         </Box>
       </StickyNavContainer>
 
@@ -148,11 +121,7 @@ export const PackageDocs: FunctionComponent = () => {
         top={TOP_OFFSET}
       >
         <Box overflowY="auto" py={4}>
-          <NavTree
-            data-event={PACKAGE_ANALYTICS.SCOPE}
-            items={sectionItems}
-            variant="sm"
-          />
+          <SecondaryDocNavigation />
         </Box>
       </StickyNavContainer>
       <NavDrawer />
