+++ src/views/Package/useIntersectingHeader/index.ts
@@ -1,0 +1,2 @@
+export { useIntersectingHeader } from "./useIntersectingHeader";
+export * from "./util";
